requirements.txt
kafka-python==2.0.6
python-dotenv==1.0.1